+++ sources/src/main/java/com/google/solutions/jitaccess/web/iap/IapPrincipal.java
@@ -19,7 +19,7 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.web.auth;
+package com.google.solutions.jitaccess.web.iap;
 
 import com.google.solutions.jitaccess.core.auth.UserId;
