+++ sources/src/test/java/com/google/solutions/jitaccess/web/TestLogAdapter.java
@@ -22,8 +22,8 @@
 package com.google.solutions.jitaccess.web;
 
 import com.google.solutions.jitaccess.core.auth.UserId;
-import com.google.solutions.jitaccess.web.auth.DeviceInfo;
-import com.google.solutions.jitaccess.web.auth.IapPrincipal;
+import com.google.solutions.jitaccess.web.iap.DeviceInfo;
+import com.google.solutions.jitaccess.web.iap.IapPrincipal;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
