+++ sources/src/test/java/com/google/solutions/jitaccess/web/iap/TestIapAssertion.java
@@ -19,7 +19,7 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.web.auth;
+package com.google.solutions.jitaccess.web.iap;
 
 import com.google.api.client.json.webtoken.JsonWebToken;
 import org.junit.jupiter.api.Test;
